+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - `XIT ACT`: Make the material group and action lists reorderable
 - `XIT SET`: Display the 12h/24h time format in the Default option
+- `XIT SORT`: Make the sorting mode list reorderable
 - `inv-compress-inventory-info`: Add a small right padding to the unload button
 
 ### Fixed